+++ app/models/policy_configuration.rb
@@ -1,0 +1,7 @@
+class PolicyConfiguration < ApplicationRecord
+  validates :policy_type, inclusion: {in: Policies.all.map(&:name)}
+
+  def policy
+    policy_type.constantize
+  end
+end